난소종양객체탐지/requirements.txt
numpy==1.24.3
opencv_python==4.8.1.78
pandas==1.5.0
Pillow==10.0.1
Pillow==10.1.0
pycocotools==2.0.6
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.3.2
tabulate==0.8.10
torch==2.1.0
tqdm==4.64.1
ultralytics==8.0.141
